Update dogcamstreamer.py
Add additional logging and some recovery state handles. Right now the AI controller is to be terminated if a disconnection is long enough, assuming that the stream is over. It might be better to pass along commands to attempt to get the stream to recover more robustly in case of a serious network disconnection.

This decision would change how we would run the program.